improve stroke precision
darcs-hash:20080810181010-485ad-196ab2a9fb103ecdf14b88e5765ae066803b7710.gz

+++ trace.h
@@ -23,7 +23,7 @@
 
 class Trace {
 public:
-	struct Point { int x; int y; };
+	struct Point { float x; float y; };
 private:
 	Point last;
 	bool active;
